ATR-855: folder restructure, keep event data in one place
