+++ src/Acuminator/Acuminator.Utilities/Roslyn/Semantic/SemanticModel/GraphEvents/Model/EventHandlerSignatureType.cs
