+++ src/Acuminator/Acuminator.Utilities/Roslyn/Semantic/SemanticModel/GraphEvents/Model/EventType.cs
