+++ src/Acuminator/Acuminator.Utilities/Roslyn/Semantic/SemanticModel/GraphEvents/Model/EventsRecognition.cs
@@ -3,11 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 
 using Acuminator.Utilities.Common;
-using Acuminator.Utilities.Roslyn.Semantic.Symbols;
 
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Microsoft.CodeAnalysis.Text;
 
 namespace Acuminator.Utilities.Roslyn.Semantic.GraphEvents
 {
@@ -36,31 +33,31 @@ namespace Acuminator.Utilities.Roslyn.Semantic.GraphEvents
 			symbol.ThrowOnNull();
 			pxContext.ThrowOnNull();
 
-			if (symbol.ReturnsVoid && symbol.TypeParameters.IsEmpty && !symbol.Parameters.IsEmpty)
+			if (!symbol.ReturnsVoid || !symbol.TypeParameters.IsDefaultOrEmpty || symbol.Parameters.IsDefaultOrEmpty)
+				return EventInfo.None(symbol);
+
+			// Loosely check method signature because sometimes business logic 
+			// is extracted from event handler calls to a separate method
+
+			// Old non-generic syntax
+			if (symbol.Parameters[0].Type.OriginalDefinition.InheritsFromOrEquals(pxContext.PXCache.Type))
 			{
-				// Loosely check method signature because sometimes business logic 
-				// is extracted from event handler calls to a separate method
+				if (symbol.Name.EndsWith("CacheAttached", StringComparison.Ordinal))
+					return new EventInfo(EventType.CacheAttached, EventHandlerSignatureType.Default);
 
-				// Old non-generic syntax
-				if (symbol.Parameters[0].Type.OriginalDefinition.InheritsFromOrEquals(pxContext.PXCache.Type))
+				if (symbol.Parameters.Length >= 2 && pxContext.Events.EventTypeMap.TryGetValue(
+						symbol.Parameters[1].Type.OriginalDefinition, out EventType eventType))
 				{
-					if (symbol.Name.EndsWith("CacheAttached", StringComparison.Ordinal))
-						return new EventInfo(EventType.CacheAttached, EventHandlerSignatureType.Default);
-
-					if (symbol.Parameters.Length >= 2 && pxContext.Events.EventTypeMap.TryGetValue(
-						    symbol.Parameters[1].Type.OriginalDefinition, out EventType eventType))
-					{
-						return new EventInfo(eventType, EventHandlerSignatureType.Default);
-					}
-				}
-				else if (pxContext.Events.EventTypeMap.TryGetValue(
-					symbol.Parameters[0].Type.OriginalDefinition, out EventType eventType)) // New generic event handler syntax
-				{
-					return new EventInfo(eventType, EventHandlerSignatureType.Generic);
+					return new EventInfo(eventType, EventHandlerSignatureType.Default);
 				}
 			}
+			else if (pxContext.Events.EventTypeMap.TryGetValue(
+				symbol.Parameters[0].Type.OriginalDefinition, out EventType eventType)) // New generic event handler syntax
+			{
+				return new EventInfo(eventType, EventHandlerSignatureType.Generic);
+			}
 
-			return EventInfo.None();
+			return EventInfo.None(symbol);
 		}
 	}
 }
